use HttpHeaders as bean
