+++ src/main/java/com/khcho902/gangnamspringbootstudy/config/RootConfig.java
@@ -1,0 +1,31 @@
+package com.khcho902.gangnamspringbootstudy.config;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpHeaders;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class RootConfig {
+
+    @Value("${naver.openapi.clientId}")
+    private String naverOpenApiClientId;
+    @Value("${naver.openapi.clientSecret}")
+    private String naverOpenApiClientSecret;
+
+    @Bean
+    public RestTemplate restTemplate() {
+        return new RestTemplate();
+    }
+
+    @Bean
+    public HttpHeaders httpHeaders(){
+        HttpHeaders httpHeaders = new HttpHeaders();
+
+        httpHeaders.add("X-Naver-Client-Id", naverOpenApiClientId);
+        httpHeaders.add("X-Naver-Client-Secret", naverOpenApiClientSecret);
+
+        return httpHeaders;
+    }
+}
