+++ src/main/java/com/khcho902/gangnamspringbootstudy/repository/BlogRepository.java
@@ -15,18 +15,11 @@ public class BlogRepository {
 
     @Value("${naver.openapi.blogUrl}")
     private String naverOpenApiUrl;
-    @Value("${naver.openapi.clientId}")
-    private String naverOpenApiClientId;
-    @Value("${naver.openapi.clientSecret}")
-    private String naverOpenApiClientSecret;
 
     private final RestTemplate restTemplate;
+    private final HttpHeaders httpHeaders;
 
     public ResponseBlog findByQuery(String query) {
-
-        HttpHeaders httpHeaders = new HttpHeaders();
-        httpHeaders.add("X-Naver-Client-Id", naverOpenApiClientId);
-        httpHeaders.add("X-Naver-Client-Secret", naverOpenApiClientSecret);
 
         String url = naverOpenApiUrl + "?query=" + query;
