--- src/main/java/com/khcho902/gangnamspringbootstudy/repository/RestTemplateConfig.java
@@ -1,14 +1,0 @@
-package com.khcho902.gangnamspringbootstudy.repository;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class RestTemplateConfig {
-
-    @Bean
-    public RestTemplate restTemplate() {
-        return new RestTemplate();
-    }
-}
